Conflicts:
	hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/blockmanagement/BlockManager.java
	hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/blockmanagement/PendingReplicationBlocks.java
	hadoop-hdfs-project/hadoop-hdfs/src/main/java/se/sics/hop/metadata/LeaderElection.java
	hadoop-hdfs-project/hadoop-hdfs/src/main/java/se/sics/hop/metadata/lock/HDFSTransactionLockAcquirer.java
	hadoop-hdfs-project/hadoop-hdfs/src/test/java/org/apache/hadoop/hdfs/server/blockmanagement/TestPendingReplication.java

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/INodeFileUnderConstruction.java
@@ -75,8 +75,8 @@ public class INodeFileUnderConstruction extends INodeFile implements MutableBloc
                              String clientName,
                              String clientMachine,
                              DatanodeID clientNode,
-                             long inodeId,
-                             long pid) {
+                             int inodeId,
+                             int pid) {
     super(perm, blocks, blockReplication, modificationTime, modificationTime,
           preferredBlockSize);
     setLocalNameNoPersistance(name);
